Modified screenshot test to add end to end integration test
