+++ gradle/libs.versions.toml
@@ -54,6 +54,7 @@ coil-compose = { group = "io.coil-kt.coil3", name = "coil-compose", version.ref 
 coil-okhttp = { group = "io.coil-kt.coil3", name = "coil-network-okhttp", version.ref = "coil" }
 room-ktx = { module = "androidx.room:room-ktx", version.ref = "room" }
 room-compiler = { module = "androidx.room:room-compiler", version.ref = "room" }
+room-testing = { module = "androidx.room:room-testing", version.ref = "room" }
 androidx-ui-tooling = { group = "androidx.compose.ui", name = "ui-tooling", version.ref = "uiTooling" }
 test-junit4 = { module = "androidx.compose.ui:ui-test-junit4", version.ref = "composeJunit" }
 test-manifest = { module = "androidx.compose.ui:ui-test-manifest", version.ref = "composeJunit" }
